Changed exception catch to pre-check in close_db()
When running the debugger and having enabled to catch raised exceptions, the AttributeError raised for every request that doesn't use the DB made it basically impossible to navigate the UI with it enabled because it would stop on the raised exception 10 times. With this change it works exactly the same way but the exception isn't raised anymore, making UI navigation way easier.