Removed check for VF being subfolder when adding
We would check whether the volume folder of the volume that will be added is a subfolder or parentfolder of the volume folder of a different volume. That is now removed, as multiple volumes having the same volume folder is allowed and thus a being a parent or child of one should also be allowed. The problems are the same.